feat: update supported Ruby versions to 3.3.10, 3.4.8, 4.0.1
- Update rubocop to 1.85.0, fix new offenses
- Split sample_jobs.rb into individual files

[BANK-1738]
