+++ .circleci/config.yml
@@ -12,9 +12,9 @@ workflows:
           matrix:
             parameters:
               ruby_version:
-                - '3.4.1'
-                - '3.3.6'
-                - '3.2.5'
+                - '4.0.1'
+                - '3.4.8'
+                - '3.3.10'
           after-appraisal-install-steps:
             - gem-tool/wait_mysql
             - run: bundle exec rubocop
