+++ .ruby-version
@@ -1,1 +1,1 @@
-ruby-3.4.1
+ruby-4.0.1
