+++ Gemfile
@@ -4,16 +4,16 @@ source 'https://rubygems.org' # global source
 
 source 'https://rubygems.org' do
   gem 'appraisal', '>= 2.5', '< 3'
-  gem 'bundler', '>= 2.6', '< 3'
+  gem 'bundler', '>= 2.6', '< 5'
   gem 'debug', '>= 1.11', '< 2'
   gem 'mysql2', '>= 0.5.6', '< 1'
   gem 'rails', '>= 7.2', '< 8.2'
-  gem 'rake', '>= 13.2', '< 14'
+  gem 'rake', '>= 13.3', '< 14'
   gem 'rspec', '>= 3.13', '< 4'
   gem 'rspec_junit_formatter', '>= 0.6', '< 1'
   gem 'rspec-rails', '>= 4', '< 5'
-  gem 'rubocop', '= 1.74'
-  gem 'rubocop-rspec', '= 3.5'
+  gem 'rubocop', '1.85.0'
+  gem 'rubocop-rspec', '3.5.0'
   gem 'shoulda-matchers', '>= 6.4', '< 7'
   gem 'simplecov', '>= 0.22', '< 1', group: :test, require: false
 end
