+++ LICENSE.txt
@@ -1,4 +1,4 @@
-Copyright (c) 2020-2025 AppFolio, Inc
+Copyright (c) 2020-2026 AppFolio, Inc
 
 Permission is hereby granted, free of charge, to any person obtaining
 a copy of this software and associated documentation files (the
