+++ spec/sample_jobs/always_fail_job.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AlwaysFailJob < ApplicationJob
+  def perform(_params)
+    raise StandardError, 'some_error'
+  end
+end
