+++ spec/sample_jobs/always_pass_job.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AlwaysPassJob < ApplicationJob
+  def perform(_params)
+    1
+  end
+end
