+++ spec/sample_jobs/application_job.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class ApplicationJob < ActiveJob::Base
+  include AeActiveJobState::HasJobState
+end
