+++ spec/sample_jobs/has_progress_job.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class HasProgressJob < ApplicationJob
+  def perform(_params)
+    set_job_progress!('percent' => 20)
+  end
+end
