+++ spec/sample_jobs/has_result_job.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class HasResultJob < ApplicationJob
+  def perform(_params)
+    set_job_result!('value' => 20)
+  end
+end
